perf: widen the header indices to usize
Review asked for this twice and was right twice. The narrow fields kept
Header at 40 bytes rather than 48, and paid for it with three checked
usize to u32 conversions on the spawn path.

Measured, since the 40 bytes was asserted rather than checked: the wider
header is faster. Spawn and await goes 27.20ns to 26.05ns, detached
26.34 to 25.32, and a 512 byte capture 31.92 to 29.67, with
non-overlapping intervals. Across executors it is the same or better at
every count, 26.2 against 27.9 at one and 54.4 against 56.3 at
sixty-four, so there is no cache penalty at scale either.

The conversions cost more than the eight bytes saved, and they could
panic. Both are gone.

+++ glommio/src/task/lifecycle_tests.rs
@@ -28,8 +28,8 @@ mod test {
         task::{Context, Poll, RawWaker, RawWakerVTable, Waker},
     };
 
-    const EXECUTOR_ID: u32 = 0;
-    const QUEUE_INDEX: u32 = 0;
+    const EXECUTOR_ID: usize = 0;
+    const QUEUE_INDEX: usize = 0;
 
     /// A waker that does nothing, for polling a `JoinHandle` whose task has
     /// already been run to completion.
@@ -57,7 +57,7 @@ mod test {
     /// take the owning-thread branch. Must be held for as long as any task from
     /// `spawn_capturing` is alive.
     fn own_tasks() -> test_executor_id::Guard {
-        test_executor_id::scoped(EXECUTOR_ID as usize)
+        test_executor_id::scoped(EXECUTOR_ID)
     }
 
     /// Spawn with a schedule closure that captures, so `RawTask::schedule`
